部分对象字段从 int 改成 long

See merge request !19

+++ pingpp/Models/Order.cs
@@ -85,7 +85,7 @@ namespace Pingpp.Models
         public Dictionary<string,object> ChargeEssentials { get; set; }
 
         [JsonProperty("available_balance")]
-        public int AvailableBalance { set; get; }
+        public long AvailableBalance { set; get; }
 
         [JsonProperty("receipt_app")]
         public string ReceiptApp { get; set; }
